Validator for actions/constructors/handlers and handler building. (#22)
Add validation for actions/constructors/handlers.
Add logic to Build the handlers after configuration and storing it.
Add logic to dedupe the instances for the purpose of configuring types.
This PR is mainly for the config path. It has Todos to finish the following unimplemented parts

TODO : Do the actual type inference and dispatching to adapters for Type configuration.